// src/api/model/notificationResponse.ts
/**
 * Generated by orval v8.12.1 🍺
 * Do not edit manually.
 * DunbarHorizon API 명세서
 * DunbarHorizon 백엔드 API
 * OpenAPI spec version: 1.0.0
 */
import type { NotificationResponseMetadata } from './notificationResponseMetadata';
import type { NotificationResponseType } from './notificationResponseType';

export interface NotificationResponse {
  id?: string;
  title?: string;
  content?: string;
  metadata?: NotificationResponseMetadata;
  type?: NotificationResponseType;
  isRead?: boolean;
  createdAt?: string;
}